+++ src/motion/features/animation.ts
@@ -14,9 +14,7 @@ const AnimationState = makeRenderlessComponent((props: FeatureProps) => {
      * to the underlying animation library. We only want to load that if we load this,
      * so people can optionally code split it out using the `m` component.
      */
-    if (!visualElement.animationState) {
-        visualElement.animationState = createAnimationState(visualElement)
-    }
+    visualElement.animationState ||= createAnimationState(visualElement)
 
     const variantContext = useVariantContext()